Pulling added, minor changes
Puller agent added
Other utilities function added (isVisible...)
Collision func added to agents

+++ src/Game/GameLauncher.java
@@ -1,5 +1,8 @@
 package Game;
 
+/** Launches a game with given parameters
+ * @author Clément
+ */
 public class GameLauncher {
 
 	public static void main(String[] args) {
@@ -7,11 +10,11 @@ public class GameLauncher {
 		int nbGoodGuys = 50;
 		int nbBadGuys = 5;
 		int agentSize = 15;
-		float minVelocity = 3;
-		float maxVelocityGG = 4;
+		float minVelocity = (float) 3;
+		float maxVelocityGG = (float) 4;
 		float maxVelocityBG = (float) 3;
-		float radius = 50;
-		float angle = 100;
+		float radius = (float) 50;
+		float angle = (float) 100;
 		// =============================================================================================================
 		// Creates the game
 		Game game = new Game(nbGoodGuys, nbBadGuys, agentSize, minVelocity,
